Massively reorganize rendergraph
